+++ openml/extensions/__init__.py
@@ -1,8 +1,10 @@
+from typing import List, Type  # noqa: F401
+
 from .extension_interface import Extension
 from .functions import register_extension, get_extension_by_model, get_extension_by_flow
 
 
-extensions = []
+extensions = []  # type: List[Type[Extension]]
 
 
 __all__ = [
