+++ tests/test_extensions/test_sklearn_extension/test_sklearn_flow_functions.py
@@ -29,13 +29,11 @@ else:
     from sklearn.impute import SimpleImputer as Imputer
 
 import openml
-from openml.extensions.sklearn.flow_functions import sklearn_to_flow, flow_to_sklearn
-from openml.testing import TestBase
+from openml.extensions.sklearn import SklearnExtension
+from openml.exceptions import PyOpenMLError
 from openml.flows import OpenMLFlow
 from openml.flows.functions import assert_flows_equal
-from openml.extensions.sklearn.flow_functions import _format_external_version, \
-    _check_dependencies, check_n_jobs
-from openml.exceptions import PyOpenMLError
+from openml.testing import TestBase
 
 this_directory = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(this_directory)
@@ -54,7 +52,7 @@ class Model(sklearn.base.BaseEstimator):
         pass
 
 
-class TestFunctions(TestBase):
+class TestSklearnExtensionFlowFunctions(TestBase):
     # Splitting not helpful, these test's don't rely on the server and take less
     # than 1 seconds
 
@@ -64,137 +62,135 @@ class TestFunctions(TestBase):
         self.X = iris.data
         self.y = iris.target
 
-    @mock.patch('openml.extensions.sklearn.flow_functions._check_dependencies')
-    def test_serialize_model(self, check_dependencies_mock):
-        model = sklearn.tree.DecisionTreeClassifier(criterion='entropy',
-                                                    max_features='auto',
-                                                    max_leaf_nodes=2000)
+        self.extension = SklearnExtension()
 
-        fixture_name = 'sklearn.tree.tree.DecisionTreeClassifier'
-        fixture_description = 'Automatically created scikit-learn flow.'
-        version_fixture = 'sklearn==%s\nnumpy>=1.6.1\nscipy>=0.9' \
-                          % sklearn.__version__
-        # min_impurity_decrease has been introduced in 0.20
-        # min_impurity_split has been deprecated in 0.20
-        if LooseVersion(sklearn.__version__) < "0.19":
-            fixture_parameters = \
-                OrderedDict((('class_weight', 'null'),
-                            ('criterion', '"entropy"'),
-                            ('max_depth', 'null'),
-                            ('max_features', '"auto"'),
-                            ('max_leaf_nodes', '2000'),
-                            ('min_impurity_split', '1e-07'),
-                            ('min_samples_leaf', '1'),
-                            ('min_samples_split', '2'),
-                            ('min_weight_fraction_leaf', '0.0'),
-                            ('presort', 'false'),
-                            ('random_state', 'null'),
-                            ('splitter', '"best"')))
-        else:
-            fixture_parameters = \
-                OrderedDict((('class_weight', 'null'),
-                            ('criterion', '"entropy"'),
-                            ('max_depth', 'null'),
-                            ('max_features', '"auto"'),
-                            ('max_leaf_nodes', '2000'),
-                            ('min_impurity_decrease', '0.0'),
-                            ('min_impurity_split', 'null'),
-                            ('min_samples_leaf', '1'),
-                            ('min_samples_split', '2'),
-                            ('min_weight_fraction_leaf', '0.0'),
-                            ('presort', 'false'),
-                            ('random_state', 'null'),
-                            ('splitter', '"best"')))
-        structure_fixture = {'sklearn.tree.tree.DecisionTreeClassifier': []}
+    def test_serialize_model(self):
+        with mock.patch.object(self.extension, '_check_dependencies') as check_dependencies_mock:
+            model = sklearn.tree.DecisionTreeClassifier(criterion='entropy',
+                                                        max_features='auto',
+                                                        max_leaf_nodes=2000)
 
-        serialization = sklearn_to_flow(model)
-        structure = serialization.get_structure('name')
+            fixture_name = 'sklearn.tree.tree.DecisionTreeClassifier'
+            fixture_description = 'Automatically created scikit-learn flow.'
+            version_fixture = 'sklearn==%s\nnumpy>=1.6.1\nscipy>=0.9' \
+                              % sklearn.__version__
+            # min_impurity_decrease has been introduced in 0.20
+            # min_impurity_split has been deprecated in 0.20
+            if LooseVersion(sklearn.__version__) < "0.19":
+                fixture_parameters = \
+                    OrderedDict((('class_weight', 'null'),
+                                ('criterion', '"entropy"'),
+                                ('max_depth', 'null'),
+                                ('max_features', '"auto"'),
+                                ('max_leaf_nodes', '2000'),
+                                ('min_impurity_split', '1e-07'),
+                                ('min_samples_leaf', '1'),
+                                ('min_samples_split', '2'),
+                                ('min_weight_fraction_leaf', '0.0'),
+                                ('presort', 'false'),
+                                ('random_state', 'null'),
+                                ('splitter', '"best"')))
+            else:
+                fixture_parameters = \
+                    OrderedDict((('class_weight', 'null'),
+                                ('criterion', '"entropy"'),
+                                ('max_depth', 'null'),
+                                ('max_features', '"auto"'),
+                                ('max_leaf_nodes', '2000'),
+                                ('min_impurity_decrease', '0.0'),
+                                ('min_impurity_split', 'null'),
+                                ('min_samples_leaf', '1'),
+                                ('min_samples_split', '2'),
+                                ('min_weight_fraction_leaf', '0.0'),
+                                ('presort', 'false'),
+                                ('random_state', 'null'),
+                                ('splitter', '"best"')))
+            structure_fixture = {'sklearn.tree.tree.DecisionTreeClassifier': []}
 
-        self.assertEqual(serialization.name, fixture_name)
-        self.assertEqual(serialization.class_name, fixture_name)
-        self.assertEqual(serialization.description, fixture_description)
-        self.assertEqual(serialization.parameters, fixture_parameters)
-        self.assertEqual(serialization.dependencies, version_fixture)
-        self.assertDictEqual(structure, structure_fixture)
+            serialization = self.extension.model_to_flow(model)
+            structure = serialization.get_structure('name')
 
-        new_model = flow_to_sklearn(serialization)
-        # compares string representations of the dict, as it potentially
-        # contains complex objects that can not be compared with == op
-        # Only in Python 3.x, as Python 2 has Unicode issues
-        if sys.version_info[0] >= 3:
-            self.assertEqual(str(model.get_params()),
-                             str(new_model.get_params()))
+            self.assertEqual(serialization.name, fixture_name)
+            self.assertEqual(serialization.class_name, fixture_name)
+            self.assertEqual(serialization.description, fixture_description)
+            self.assertEqual(serialization.parameters, fixture_parameters)
+            self.assertEqual(serialization.dependencies, version_fixture)
+            self.assertDictEqual(structure, structure_fixture)
 
-        self.assertEqual(type(new_model), type(model))
-        self.assertIsNot(new_model, model)
+            new_model = self.extension.flow_to_model(serialization)
+            # compares string representations of the dict, as it potentially
+            # contains complex objects that can not be compared with == op
+            # Only in Python 3.x, as Python 2 has Unicode issues
+            if sys.version_info[0] >= 3:
+                self.assertEqual(str(model.get_params()), str(new_model.get_params()))
 
-        self.assertEqual(new_model.get_params(), model.get_params())
-        new_model.fit(self.X, self.y)
+            self.assertEqual(type(new_model), type(model))
+            self.assertIsNot(new_model, model)
 
-        self.assertEqual(check_dependencies_mock.call_count, 1)
+            self.assertEqual(new_model.get_params(), model.get_params())
+            new_model.fit(self.X, self.y)
 
-    @mock.patch('openml.extensions.sklearn.flow_functions._check_dependencies')
-    def test_serialize_model_clustering(self, check_dependencies_mock):
-        model = sklearn.cluster.KMeans()
+            self.assertEqual(check_dependencies_mock.call_count, 1)
 
-        fixture_name = 'sklearn.cluster.k_means_.KMeans'
-        fixture_description = 'Automatically created scikit-learn flow.'
-        version_fixture = 'sklearn==%s\nnumpy>=1.6.1\nscipy>=0.9' \
-                          % sklearn.__version__
-        # n_jobs default has changed to None in 0.20
-        if LooseVersion(sklearn.__version__) < "0.20":
-            fixture_parameters = \
-                OrderedDict((('algorithm', '"auto"'),
-                             ('copy_x', 'true'),
-                             ('init', '"k-means++"'),
-                             ('max_iter', '300'),
-                             ('n_clusters', '8'),
-                             ('n_init', '10'),
-                             ('n_jobs', '1'),
-                             ('precompute_distances', '"auto"'),
-                             ('random_state', 'null'),
-                             ('tol', '0.0001'),
-                             ('verbose', '0')))
-        else:
-            fixture_parameters = \
-                OrderedDict((('algorithm', '"auto"'),
-                             ('copy_x', 'true'),
-                             ('init', '"k-means++"'),
-                             ('max_iter', '300'),
-                             ('n_clusters', '8'),
-                             ('n_init', '10'),
-                             ('n_jobs', 'null'),
-                             ('precompute_distances', '"auto"'),
-                             ('random_state', 'null'),
-                             ('tol', '0.0001'),
-                             ('verbose', '0')))
-        fixture_structure = {'sklearn.cluster.k_means_.KMeans': []}
+    def test_serialize_model_clustering(self):
+        with mock.patch.object(self.extension, '_check_dependencies') as check_dependencies_mock:
+            model = sklearn.cluster.KMeans()
 
-        serialization = sklearn_to_flow(model)
-        structure = serialization.get_structure('name')
+            fixture_name = 'sklearn.cluster.k_means_.KMeans'
+            fixture_description = 'Automatically created scikit-learn flow.'
+            version_fixture = 'sklearn==%s\nnumpy>=1.6.1\nscipy>=0.9' \
+                              % sklearn.__version__
+            # n_jobs default has changed to None in 0.20
+            if LooseVersion(sklearn.__version__) < "0.20":
+                fixture_parameters = \
+                    OrderedDict((('algorithm', '"auto"'),
+                                 ('copy_x', 'true'),
+                                 ('init', '"k-means++"'),
+                                 ('max_iter', '300'),
+                                 ('n_clusters', '8'),
+                                 ('n_init', '10'),
+                                 ('n_jobs', '1'),
+                                 ('precompute_distances', '"auto"'),
+                                 ('random_state', 'null'),
+                                 ('tol', '0.0001'),
+                                 ('verbose', '0')))
+            else:
+                fixture_parameters = \
+                    OrderedDict((('algorithm', '"auto"'),
+                                 ('copy_x', 'true'),
+                                 ('init', '"k-means++"'),
+                                 ('max_iter', '300'),
+                                 ('n_clusters', '8'),
+                                 ('n_init', '10'),
+                                 ('n_jobs', 'null'),
+                                 ('precompute_distances', '"auto"'),
+                                 ('random_state', 'null'),
+                                 ('tol', '0.0001'),
+                                 ('verbose', '0')))
+            fixture_structure = {'sklearn.cluster.k_means_.KMeans': []}
 
-        self.assertEqual(serialization.name, fixture_name)
-        self.assertEqual(serialization.class_name, fixture_name)
-        self.assertEqual(serialization.description, fixture_description)
-        self.assertEqual(serialization.parameters, fixture_parameters)
-        self.assertEqual(serialization.dependencies, version_fixture)
-        self.assertDictEqual(structure, fixture_structure)
+            serialization = self.extension.model_to_flow(model)
+            structure = serialization.get_structure('name')
 
-        new_model = flow_to_sklearn(serialization)
-        # compares string representations of the dict, as it potentially
-        # contains complex objects that can not be compared with == op
-        # Only in Python 3.x, as Python 2 has Unicode issues
-        if sys.version_info[0] >= 3:
-            self.assertEqual(str(model.get_params()),
-                             str(new_model.get_params()))
+            self.assertEqual(serialization.name, fixture_name)
+            self.assertEqual(serialization.class_name, fixture_name)
+            self.assertEqual(serialization.description, fixture_description)
+            self.assertEqual(serialization.parameters, fixture_parameters)
+            self.assertEqual(serialization.dependencies, version_fixture)
+            self.assertDictEqual(structure, fixture_structure)
 
-        self.assertEqual(type(new_model), type(model))
-        self.assertIsNot(new_model, model)
+            new_model = self.extension.flow_to_model(serialization)
+            # compares string representations of the dict, as it potentially
+            # contains complex objects that can not be compared with == op
+            self.assertEqual(str(model.get_params()), str(new_model.get_params()))
 
-        self.assertEqual(new_model.get_params(), model.get_params())
-        new_model.fit(self.X)
+            self.assertEqual(type(new_model), type(model))
+            self.assertIsNot(new_model, model)
 
-        self.assertEqual(check_dependencies_mock.call_count, 1)
+            self.assertEqual(new_model.get_params(), model.get_params())
+            new_model.fit(self.X)
+
+            self.assertEqual(check_dependencies_mock.call_count, 1)
 
     def test_serialize_model_with_subcomponent(self):
         model = sklearn.ensemble.AdaBoostClassifier(
@@ -212,7 +208,7 @@ class TestFunctions(TestBase):
             'sklearn.tree.tree.DecisionTreeClassifier': ['base_estimator']
         }
 
-        serialization = sklearn_to_flow(model)
+        serialization = self.extension.model_to_flow(model)
         structure = serialization.get_structure('name')
 
         self.assertEqual(serialization.name, fixture_name)
@@ -230,13 +226,10 @@ class TestFunctions(TestBase):
                          fixture_subcomponent_description)
         self.assertDictEqual(structure, fixture_structure)
 
-        new_model = flow_to_sklearn(serialization)
+        new_model = self.extension.flow_to_model(serialization)
         # compares string representations of the dict, as it potentially
         # contains complex objects that can not be compared with == op
-        # Only in Python 3.x, as Python 2 has Unicode issues
-        if sys.version_info[0] >= 3:
-            self.assertEqual(str(model.get_params()),
-                             str(new_model.get_params()))
+        self.assertEqual(str(model.get_params()), str(new_model.get_params()))
 
         self.assertEqual(type(new_model), type(model))
         self.assertIsNot(new_model, model)
@@ -268,7 +261,7 @@ class TestFunctions(TestBase):
             'sklearn.dummy.DummyClassifier': ['dummy']
         }
 
-        serialization = sklearn_to_flow(model)
+        serialization = self.extension.model_to_flow(model)
         structure = serialization.get_structure('name')
 
         self.assertEqual(serialization.name, fixture_name)
@@ -308,7 +301,7 @@ class TestFunctions(TestBase):
         self.assertIsInstance(serialization.components['dummy'],
                               OpenMLFlow)
 
-        new_model = flow_to_sklearn(serialization)
+        new_model = self.extension.flow_to_model(serialization)
         # compares string representations of the dict, as it potentially
         # contains complex objects that can not be compared with == op
         # Only in Python 3.x, as Python 2 has Unicode issues
@@ -352,7 +345,7 @@ class TestFunctions(TestBase):
             'sklearn.cluster.k_means_.KMeans': ['clusterer']
         }
 
-        serialization = sklearn_to_flow(model)
+        serialization = self.extension.model_to_flow(model)
         structure = serialization.get_structure('name')
 
         self.assertEqual(serialization.name, fixture_name)
@@ -391,7 +384,7 @@ class TestFunctions(TestBase):
                               OpenMLFlow)
 
         # del serialization.model
-        new_model = flow_to_sklearn(serialization)
+        new_model = self.extension.flow_to_model(serialization)
         # compares string representations of the dict, as it potentially
         # contains complex objects that can not be compared with == op
         # Only in Python 3.x, as Python 2 has Unicode issues
@@ -440,13 +433,13 @@ class TestFunctions(TestBase):
             'sklearn.preprocessing._encoders.OneHotEncoder': ['nominal']
         }
 
-        serialization = sklearn_to_flow(model)
+        serialization = self.extension.model_to_flow(model)
         structure = serialization.get_structure('name')
         self.assertEqual(serialization.name, fixture)
         self.assertEqual(serialization.description, fixture_description)
         self.assertDictEqual(structure, fixture_structure)
         # del serialization.model
-        new_model = flow_to_sklearn(serialization)
+        new_model = self.extension.flow_to_model(serialization)
         # compares string representations of the dict, as it potentially
         # contains complex objects that can not be compared with == op
         # Only in Python 3.x, as Python 2 has Unicode issues
@@ -455,7 +448,7 @@ class TestFunctions(TestBase):
                              str(new_model.get_params()))
         self.assertEqual(type(new_model), type(model))
         self.assertIsNot(new_model, model)
-        serialization2 = sklearn_to_flow(new_model)
+        serialization2 = self.extension.model_to_flow(new_model)
         assert_flows_equal(serialization, serialization2)
 
     @unittest.skipIf(LooseVersion(sklearn.__version__) < "0.20",
@@ -492,19 +485,19 @@ class TestFunctions(TestBase):
         }
 
         fixture_description = 'Automatically created scikit-learn flow.'
-        serialization = sklearn_to_flow(model)
+        serialization = self.extension.model_to_flow(model)
         structure = serialization.get_structure('name')
         self.assertEqual(serialization.name, fixture_name)
         self.assertEqual(serialization.description, fixture_description)
         self.assertDictEqual(structure, fixture_structure)
         # del serialization.model
-        new_model = flow_to_sklearn(serialization)
+        new_model = self.extension.flow_to_model(serialization)
         # compares string representations of the dict, as it potentially
         # contains complex objects that can not be compared with == op
         self.assertEqual(str(model.get_params()), str(new_model.get_params()))
         self.assertEqual(type(new_model), type(model))
         self.assertIsNot(new_model, model)
-        serialization2 = sklearn_to_flow(new_model)
+        serialization2 = self.extension.model_to_flow(new_model)
         assert_flows_equal(serialization, serialization2)
 
     def test_serialize_feature_union(self):
@@ -515,8 +508,9 @@ class TestFunctions(TestBase):
         scaler = sklearn.preprocessing.StandardScaler()
 
         fu = sklearn.pipeline.FeatureUnion(
-            transformer_list=[('ohe', ohe), ('scaler', scaler)])
-        serialization = sklearn_to_flow(fu)
+            transformer_list=[('ohe', ohe), ('scaler', scaler)]
+        )
+        serialization = self.extension.model_to_flow(fu)
         structure = serialization.get_structure('name')
         # OneHotEncoder was moved to _encoders module in 0.20
         module_name_encoder = ('_encoders'
@@ -534,7 +528,7 @@ class TestFunctions(TestBase):
         }
         self.assertEqual(serialization.name, fixture_name)
         self.assertDictEqual(structure, fixture_structure)
-        new_model = flow_to_sklearn(serialization)
+        new_model = self.extension.flow_to_model(serialization)
         # compares string representations of the dict, as it potentially
         # contains complex objects that can not be compared with == op
         # Only in Python 3.x, as Python 2 has Unicode issues
@@ -573,12 +567,12 @@ class TestFunctions(TestBase):
         new_model.fit(self.X, self.y)
 
         fu.set_params(scaler=None)
-        serialization = sklearn_to_flow(fu)
+        serialization = self.extension.model_to_flow(fu)
         self.assertEqual(serialization.name,
                          'sklearn.pipeline.FeatureUnion('
                          'ohe=sklearn.preprocessing.{}.OneHotEncoder)'
                          .format(module_name_encoder))
-        new_model = flow_to_sklearn(serialization)
+        new_model = self.extension.flow_to_model(serialization)
         self.assertEqual(type(new_model), type(fu))
         self.assertIsNot(new_model, fu)
         self.assertIs(new_model.transformer_list[1][1], None)
@@ -592,8 +586,8 @@ class TestFunctions(TestBase):
             transformer_list=[('ohe', ohe), ('scaler', scaler)])
         fu2 = sklearn.pipeline.FeatureUnion(
             transformer_list=[('scaler', ohe), ('ohe', scaler)])
-        fu1_serialization = sklearn_to_flow(fu1)
-        fu2_serialization = sklearn_to_flow(fu2)
+        fu1_serialization = self.extension.model_to_flow(fu1)
+        fu2_serialization = self.extension.model_to_flow(fu2)
         # OneHotEncoder was moved to _encoders module in 0.20
         module_name_encoder = ('_encoders'
                                if LooseVersion(sklearn.__version__) >= "0.20"
@@ -628,7 +622,7 @@ class TestFunctions(TestBase):
         cv = sklearn.model_selection.StratifiedKFold(n_splits=5, shuffle=True)
         rs = sklearn.model_selection.RandomizedSearchCV(
             estimator=model, param_distributions=parameter_grid, cv=cv)
-        serialized = sklearn_to_flow(rs)
+        serialized = self.extension.model_to_flow(rs)
         structure = serialized.get_structure('name')
         # OneHotEncoder was moved to _encoders module in 0.20
         module_name_encoder = ('_encoders'
@@ -656,17 +650,14 @@ class TestFunctions(TestBase):
         self.assertEqual(structure, fixture_structure)
 
         # now do deserialization
-        deserialized = flow_to_sklearn(serialized)
+        deserialized = self.extension.flow_to_model(serialized)
         # compares string representations of the dict, as it potentially
         # contains complex objects that can not be compared with == op
         # JvR: compare str length, due to memory address of distribution
-        # Only in Python 3.x, as Python 2 has Unicode issues
-        if sys.version_info[0] >= 3:
-            self.assertEqual(len(str(rs.get_params())),
-                             len(str(deserialized.get_params())))
+        self.assertEqual(len(str(rs.get_params())), len(str(deserialized.get_params())))
 
         # Checks that sklearn_to_flow is idempotent.
-        serialized2 = sklearn_to_flow(deserialized)
+        serialized2 = self.extension.model_to_flow(deserialized)
         self.assertNotEqual(rs, deserialized)
         # Would raise an exception if the flows would be unequal
         assert_flows_equal(serialized, serialized2)
@@ -676,8 +667,8 @@ class TestFunctions(TestBase):
                            int, np.int, np.int32, np.int64]
 
         for supported_type in supported_types:
-            serialized = sklearn_to_flow(supported_type)
-            deserialized = flow_to_sklearn(serialized)
+            serialized = self.extension.model_to_flow(supported_type)
+            deserialized = self.extension.flow_to_model(serialized)
             self.assertEqual(deserialized, supported_type)
 
     def test_serialize_rvs(self):
@@ -686,8 +677,8 @@ class TestFunctions(TestBase):
                          scipy.stats.randint(low=-3, high=15)]
 
         for supported_rv in supported_rvs:
-            serialized = sklearn_to_flow(supported_rv)
-            deserialized = flow_to_sklearn(serialized)
+            serialized = self.extension.model_to_flow(supported_rv)
+            deserialized = self.extension.flow_to_model(serialized)
             self.assertEqual(type(deserialized.dist), type(supported_rv.dist))
             del deserialized.dist
             del supported_rv.dist
@@ -695,8 +686,8 @@ class TestFunctions(TestBase):
                              supported_rv.__dict__)
 
     def test_serialize_function(self):
-        serialized = sklearn_to_flow(sklearn.feature_selection.chi2)
-        deserialized = flow_to_sklearn(serialized)
+        serialized = self.extension.model_to_flow(sklearn.feature_selection.chi2)
+        deserialized = self.extension.flow_to_model(serialized)
         self.assertEqual(deserialized, sklearn.feature_selection.chi2)
 
     def test_serialize_cvobject(self):
@@ -723,10 +714,10 @@ class TestFunctions(TestBase):
             ]),
         ]
         for method, fixture in zip(methods, fixtures):
-            m = sklearn_to_flow(method)
+            m = self.extension.model_to_flow(method)
             self.assertEqual(m, fixture)
 
-            m_new = flow_to_sklearn(m)
+            m_new = self.extension.flow_to_model(m)
             self.assertIsNot(m_new, m)
             self.assertIsInstance(m_new, type(method))
 
@@ -749,8 +740,8 @@ class TestFunctions(TestBase):
               "criterion": ["gini", "entropy"]}]
 
         for grid, model in zip(grids, models):
-            serialized = sklearn_to_flow(grid)
-            deserialized = flow_to_sklearn(serialized)
+            serialized = self.extension.model_to_flow(grid)
+            deserialized = self.extension.flow_to_model(serialized)
 
             self.assertEqual(deserialized, grid)
             self.assertIsNot(deserialized, grid)
@@ -758,8 +749,8 @@ class TestFunctions(TestBase):
             hpo = sklearn.model_selection.GridSearchCV(
                 param_grid=grid, estimator=model)
 
-            serialized = sklearn_to_flow(hpo)
-            deserialized = flow_to_sklearn(serialized)
+            serialized = self.extension.model_to_flow(hpo)
+            deserialized = self.extension.flow_to_model(serialized)
             self.assertEqual(hpo.param_grid, deserialized.param_grid)
             self.assertEqual(hpo.estimator.get_params(),
                              deserialized.estimator.get_params())
@@ -790,8 +781,8 @@ class TestFunctions(TestBase):
                  'reduce_dim__k': N_FEATURES_OPTIONS,
                  'classify__C': C_OPTIONS}]
 
-        serialized = sklearn_to_flow(grid)
-        deserialized = flow_to_sklearn(serialized)
+        serialized = self.extension.model_to_flow(grid)
+        deserialized = self.extension.flow_to_model(serialized)
 
         self.assertEqual(grid[0]['reduce_dim'][0].get_params(),
                          deserialized[0]['reduce_dim'][0].get_params())
@@ -817,8 +808,8 @@ class TestFunctions(TestBase):
     def test_serialize_resampling(self):
         kfold = sklearn.model_selection.StratifiedKFold(
             n_splits=4, shuffle=True)
-        serialized = sklearn_to_flow(kfold)
-        deserialized = flow_to_sklearn(serialized)
+        serialized = self.extension.model_to_flow(kfold)
+        deserialized = self.extension.flow_to_model(serialized)
         # Best approximation to get_params()
         self.assertEqual(str(deserialized), str(kfold))
         self.assertIsNot(deserialized, kfold)
@@ -830,9 +821,9 @@ class TestFunctions(TestBase):
 
         model = Model('true', '1', '0.1')
 
-        serialized = sklearn_to_flow(model)
+        serialized = self.extension.model_to_flow(model)
         serialized.external_version = 'sklearn==test123'
-        deserialized = flow_to_sklearn(serialized)
+        deserialized = self.extension.flow_to_model(serialized)
         self.assertEqual(deserialized.get_params(), model.get_params())
         self.assertIsNot(deserialized, model)
 
@@ -841,12 +832,11 @@ class TestFunctions(TestBase):
         kernel = sklearn.gaussian_process.kernels.Matern()
         gp = sklearn.gaussian_process.GaussianProcessClassifier(
             kernel=kernel, optimizer=opt)
-        self.assertRaisesRegex(
+        with self.assertRaisesRegex(
             TypeError,
-            r"Matern\(length_scale=1, nu=1.5\), "
-            "<class 'sklearn.gaussian_process.kernels.Matern'>",
-            sklearn_to_flow, gp,
-        )
+            r"Matern\(length_scale=1, nu=1.5\), <class 'sklearn.gaussian_process.kernels.Matern'>",
+        ):
+            self.extension.model_to_flow(gp)
 
     def test_error_on_adding_component_multiple_times_to_flow(self):
         # this function implicitly checks
@@ -854,21 +844,22 @@ class TestFunctions(TestBase):
         pca = sklearn.decomposition.PCA()
         pca2 = sklearn.decomposition.PCA()
         pipeline = sklearn.pipeline.Pipeline((('pca1', pca), ('pca2', pca2)))
-        fixture = "Found a second occurence of component .*.PCA when trying " \
-                  "to serialize Pipeline"
-        self.assertRaisesRegex(ValueError, fixture, sklearn_to_flow, pipeline)
+        fixture = "Found a second occurence of component .*.PCA when trying to serialize Pipeline"
+        with self.assertRaisesRegex(ValueError, fixture):
+            self.extension.model_to_flow(pipeline)
 
         fu = sklearn.pipeline.FeatureUnion((('pca1', pca), ('pca2', pca2)))
         fixture = "Found a second occurence of component .*.PCA when trying " \
                   "to serialize FeatureUnion"
-        self.assertRaisesRegex(ValueError, fixture, sklearn_to_flow, fu)
+        with self.assertRaisesRegex(ValueError, fixture):
+            self.extension.model_to_flow(fu)
 
         fs = sklearn.feature_selection.SelectKBest()
         fu2 = sklearn.pipeline.FeatureUnion((('pca1', pca), ('fs', fs)))
         pipeline2 = sklearn.pipeline.Pipeline((('fu', fu2), ('pca2', pca2)))
-        fixture = "Found a second occurence of component .*.PCA when trying " \
-                  "to serialize Pipeline"
-        self.assertRaisesRegex(ValueError, fixture, sklearn_to_flow, pipeline2)
+        fixture = "Found a second occurence of component .*.PCA when trying to serialize Pipeline"
+        with self.assertRaisesRegex(ValueError, fixture):
+            self.extension.model_to_flow(pipeline2)
 
     def test_subflow_version_propagated(self):
         this_directory = os.path.dirname(os.path.abspath(__file__))
@@ -879,22 +870,22 @@ class TestFunctions(TestBase):
         pca = sklearn.decomposition.PCA()
         dummy = tests.test_flows.dummy_learn.dummy_forest.DummyRegressor()
         pipeline = sklearn.pipeline.Pipeline((('pca', pca), ('dummy', dummy)))
-        flow = sklearn_to_flow(pipeline)
+        flow = self.extension.model_to_flow(pipeline)
         # In python2.7, the unit tests work differently on travis-ci; therefore,
         # I put the alternative travis-ci answer here as well. While it has a
         # different value, it is still correct as it is a propagation of the
         # subclasses' module name
         self.assertEqual(flow.external_version, '%s,%s,%s' % (
-            _format_external_version('openml', openml.__version__),
-            _format_external_version('sklearn', sklearn.__version__),
-            _format_external_version('tests', '0.1')))
+            self.extension._format_external_version('openml', openml.__version__),
+            self.extension._format_external_version('sklearn', sklearn.__version__),
+            self.extension._format_external_version('tests', '0.1')))
 
     @mock.patch('warnings.warn')
     def test_check_dependencies(self, warnings_mock):
         dependencies = ['sklearn==0.1', 'sklearn>=99.99.99',
                         'sklearn>99.99.99']
         for dependency in dependencies:
-            self.assertRaises(ValueError, _check_dependencies, dependency)
+            self.assertRaises(ValueError, self.extension._check_dependencies, dependency)
 
     def test_illegal_parameter_names(self):
         # illegal name: estimators
@@ -909,7 +900,7 @@ class TestFunctions(TestBase):
         cases = [clf1, clf2]
 
         for case in cases:
-            self.assertRaises(PyOpenMLError, sklearn_to_flow, case)
+            self.assertRaises(PyOpenMLError, self.extension.model_to_flow, case)
 
     def test_illegal_parameter_names_pipeline(self):
         # illegal name: steps
@@ -971,11 +962,11 @@ class TestFunctions(TestBase):
         answers = [True, False, False, True, False, False, True, False]
 
         for model, expected_answer in zip(legal_models, answers):
-            self.assertEqual(check_n_jobs(model), expected_answer)
+            self.assertEqual(self.extension._check_n_jobs(model), expected_answer)
 
         for model in illegal_models:
             with self.assertRaises(PyOpenMLError):
-                check_n_jobs(model)
+                self.extension._check_n_jobs(model)
 
     def test__get_fn_arguments_with_defaults(self):
         if LooseVersion(sklearn.__version__) < "0.19":
@@ -993,7 +984,7 @@ class TestFunctions(TestBase):
 
         for fn, num_params_with_defaults in fns:
             defaults, defaultless = (
-                openml.extensions.sklearn.flow_functions._get_fn_arguments_with_defaults(fn)
+                self.extension._get_fn_arguments_with_defaults(fn)
             )
             self.assertIsInstance(defaults, dict)
             self.assertIsInstance(defaultless, set)
@@ -1020,15 +1011,14 @@ class TestFunctions(TestBase):
                   'OneHotEncoder__sparse': False,
                   'Estimator__min_samples_leaf': 42}
         pipe_adjusted.set_params(**params)
-        flow = openml.extensions.sklearn.flow_functions.sklearn_to_flow(pipe_adjusted)
-        pipe_deserialized = openml.extensions.sklearn.flow_functions.flow_to_sklearn(
-            flow, initialize_with_defaults=True)
+        flow = self.extension.model_to_flow(pipe_adjusted)
+        pipe_deserialized = self.extension.flow_to_model(flow, initialize_with_defaults=True)
 
         # we want to compare pipe_deserialized and pipe_orig. We use the flow
         # equals function for this
         assert_flows_equal(
-            openml.extensions.sklearn.flow_functions.sklearn_to_flow(pipe_orig),
-            openml.extensions.sklearn.flow_functions.sklearn_to_flow(pipe_deserialized),
+            self.extension.model_to_flow(pipe_orig),
+            self.extension.model_to_flow(pipe_deserialized),
         )
 
     def test_deserialize_adaboost_with_defaults(self):
@@ -1046,15 +1036,14 @@ class TestFunctions(TestBase):
                   'OneHotEncoder__sparse': False,
                   'Estimator__n_estimators': 10}
         pipe_adjusted.set_params(**params)
-        flow = openml.extensions.sklearn.flow_functions.sklearn_to_flow(pipe_adjusted)
-        pipe_deserialized = openml.extensions.sklearn.flow_functions.flow_to_sklearn(
-            flow, initialize_with_defaults=True)
+        flow = self.extension.model_to_flow(pipe_adjusted)
+        pipe_deserialized = self.extension.flow_to_model(flow, initialize_with_defaults=True)
 
         # we want to compare pipe_deserialized and pipe_orig. We use the flow
         # equals function for this
         assert_flows_equal(
-            openml.extensions.sklearn.flow_functions.sklearn_to_flow(pipe_orig),
-            openml.extensions.sklearn.flow_functions.sklearn_to_flow(pipe_deserialized),
+            self.extension.model_to_flow(pipe_orig),
+            self.extension.model_to_flow(pipe_deserialized),
         )
 
     def test_deserialize_complex_with_defaults(self):
@@ -1085,17 +1074,14 @@ class TestFunctions(TestBase):
                   'Estimator__base_estimator__base_estimator__learning_rate': 0.1,
                   'Estimator__base_estimator__base_estimator__loss__n_neighbors': 13}
         pipe_adjusted.set_params(**params)
-        flow = openml.extensions.sklearn.flow_functions.sklearn_to_flow(pipe_adjusted)
-        pipe_deserialized = openml.extensions.sklearn.flow_functions.flow_to_sklearn(
-            flow,
-            initialize_with_defaults=True,
-        )
+        flow = self.extension.model_to_flow(pipe_adjusted)
+        pipe_deserialized = self.extension.flow_to_model(flow, initialize_with_defaults=True)
 
         # we want to compare pipe_deserialized and pipe_orig. We use the flow
         # equals function for this
         assert_flows_equal(
-            openml.extensions.sklearn.flow_functions.sklearn_to_flow(pipe_orig),
-            openml.extensions.sklearn.flow_functions.sklearn_to_flow(pipe_deserialized),
+            self.extension.model_to_flow(pipe_orig),
+            self.extension.model_to_flow(pipe_deserialized),
         )
 
     def test_openml_param_name_to_sklearn(self):
@@ -1104,7 +1090,7 @@ class TestFunctions(TestBase):
             base_estimator=sklearn.tree.DecisionTreeClassifier())
         model = sklearn.pipeline.Pipeline(steps=[
             ('scaler', scaler), ('boosting', boosting)])
-        flow = openml.extensions.sklearn.flow_functions.sklearn_to_flow(model)
+        flow = self.extension.model_to_flow(model)
         task = openml.tasks.get_task(115)
         run = openml.runs.run_flow_on_task(flow, task)
         run = run.publish()
@@ -1115,8 +1101,7 @@ class TestFunctions(TestBase):
         self.assertGreater(len(setup.parameters), 15)
 
         for parameter in setup.parameters.values():
-            sklearn_name = openml.extensions.sklearn.flow_functions.openml_param_name_to_sklearn(
-                parameter, flow)
+            sklearn_name = self.extension._openml_param_name_to_sklearn(parameter, flow)
 
             # test the inverse. Currently, OpenML stores the hyperparameter
             # fullName as flow.name + flow.version + parameter.name on the
@@ -1135,30 +1120,22 @@ class TestFunctions(TestBase):
 
     def test_obtain_parameter_values_flow_not_from_server(self):
         model = sklearn.linear_model.LogisticRegression(solver='lbfgs')
-        flow = sklearn_to_flow(model)
+        flow = self.extension.model_to_flow(model)
         msg = 'Flow sklearn.linear_model.logistic.LogisticRegression has no ' \
               'flow_id!'
 
-        self.assertRaisesRegex(
-            ValueError,
-            msg,
-            openml.extensions.sklearn.flow_functions.obtain_parameter_values,
-            flow,
-        )
+        with self.assertRaisesRegex(ValueError, msg):
+            self.extension.obtain_parameter_values(flow)
 
         model = sklearn.ensemble.AdaBoostClassifier(
             base_estimator=sklearn.linear_model.LogisticRegression(
                 solver='lbfgs',
             )
         )
-        flow = sklearn_to_flow(model)
+        flow = self.extension.model_to_flow(model)
         flow.flow_id = 1
-        self.assertRaisesRegex(
-            ValueError,
-            msg,
-            openml.extensions.sklearn.flow_functions.obtain_parameter_values,
-            flow,
-        )
+        with self.assertRaisesRegex(ValueError, msg):
+            self.extension.obtain_parameter_values(flow)
 
     def test_obtain_parameter_values(self):
 
@@ -1173,10 +1150,10 @@ class TestFunctions(TestBase):
             cv=sklearn.model_selection.StratifiedKFold(n_splits=2,
                                                        random_state=1),
             n_iter=5)
-        flow = sklearn_to_flow(model)
+        flow = self.extension.model_to_flow(model)
         flow.flow_id = 1
         flow.components['estimator'].flow_id = 2
-        parameters = openml.extensions.sklearn.flow_functions.obtain_parameter_values(flow)
+        parameters = self.extension.obtain_parameter_values(flow)
         for parameter in parameters:
             self.assertIsNotNone(parameter['oml:component'], msg=parameter)
             if parameter['oml:name'] == 'n_estimators':
@@ -1189,11 +1166,10 @@ class TestFunctions(TestBase):
             max_depth=np.float64(3.0),
             min_samples_leaf=np.int32(5)
         )
-        sklearn_to_flow(dt)
+        self.extension.model_to_flow(dt)
 
     def test_numpy_array_not_allowed_in_flow(self):
         """ Simple numpy arrays should not be serializable. """
-        bin = sklearn.preprocessing.MultiLabelBinarizer(
-            classes=np.asarray([1, 2, 3])
-        )
-        self.assertRaises(TypeError, sklearn_to_flow, bin)
+        bin = sklearn.preprocessing.MultiLabelBinarizer(classes=np.asarray([1, 2, 3]))
+        with self.assertRaises(TypeError):
+            self.extension.model_to_flow(bin)
