+++ tests/test_runs/test_run.py
@@ -11,7 +11,6 @@ from sklearn.preprocessing import Imputer
 
 from openml.testing import TestBase
 import openml
-import openml.extensions.sklearn.flow_functions
 import openml.extensions.sklearn
